fix(owners): Fix ordering of related releases
This adjusts the "recent releases" behavior to use releases that are **at or before** the first release, as newer releases certainly **cannot** be responsible for the issue.

It also moves the ``get_closest_releases`` into the committers utils as its not used by any other party.
